added initial files
communication between the rio and the pico where the pico will drive a led strip. it will receive data from the rio and will output a colour/effect based on the inputted data.

currently communication is set to use the UART protocol, may use spi or other methods that may come up.